bug fixes and quality of life things added
some bugs and >>
Added UI to the buttons in the stores
Made Buy button apparent in the Customization [In the middle of the screen]
Fixed tutorial voice over bug [If changed language in the middle]
Fixed short touch Ray [when you touch and nothing happens]
Fixed Machine hitbox being too small, and moved the UI of coin on machine to be in the collision area.
Added UI Above workers to show their happiness change.